Fix for search term highlighting font size and TOC updates.
When non-lead section of article were retrieved while TOC was
open, there was a regression which caused new sections/images
to not appear in the TOC until it was closed and opened again.

Ran ttfautohint on WikiFont.ttf

Fix for sharing some articles with special characters in
their titles causing app crash.

When you select a search result the top menu switches back
to default chrome.

Change-Id: I1d96b8f9d01e5077f5091a4ad161a36272ac6112