Work in progress.
Added cocoapods and through it, AFNetworking and HPPLE.

Made Article object use AFNetworking to retrieve sections.

Fixed bug which caused lead sections to sometime jitter
halfway down the page as they downloaded. Was related to
the extra scrolling spacer only being added to the bottom
of the page after all sections had be retrieved. The fix
involved adding the last modified by, other langs, and
the spacer to bottom of lead section so it's always
there, then when non lead sections are retrieved they
are injected just above this footer stuff.

Change-Id: I991fac86a8788d3fd25595849b618a3d1ee7e699