Enhanced sharing with cards
Now, users will be able share an image card with overlaid text.
The user can still use the regular share option, which will fall
back to using an image without overlaid text, as well as a text
snippet that can be modified within the social sharing framework
provided by the system.

The user will have two options:
(1) Select text, and from a contextual menu choose to Share
(2) Select text (or don't!) and tap the upward arrow share button

For now, a fairly simple heuristic is used to identify a text
snippet in the event that the user has not selected more than the
mimimum required amount of text (currently, 10 characters). In the
future, though, the text extracts API (explaintext) could be added
to action=mobileview so it's handily available. Your code author
recommends avoiding extra roundtrips for the extract.

Change-Id: I589890b97d484230b5ddfcc774eefcf8a43cb3c3